requirements.txt
selenium
python-dotenv
openpyxl
supabase
requests